Decoupled Grid System From BirdController
Added a new class Grid which provides an interface for knowing where the lanes are currently.
GridToWorldPosition gives the position the bird will have if it is at the given grid position.
GridToForwardAxis gives the forward direction for the bird at a given grid position.
The grid coords are positive and 0-indexed from the bottom left corner.