Re-Structured Module Spawning
Previously, TrackChangeTriggers served both the purpose of signifying a turn in the path, and the place where a new module should be spawned. These functions have now been separated into TrackChangeTrigger for turns and ModuleEndNode for where to spawn a new module. Along with this change is pretty new Tiles.